allows momentjs injection
This change replaces the date parameter with a formatter parameter which
allows you to insert a custom formatter function. So you are able to
define the momentjs instance according to your needs. For example it
enables you to define the locale, enable strict mode and specify a date
format matching pattern.